pyproject.toml
[project]
name = "fetch"
dynamic = ["version"]
description = "Fetch URL to Text"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cloudscraper>=1.2.71",
    "html2text>=2025.4.15",
    "readability-lxml>=0.8.4.1",
    "beautifulsoup4>=4.12.0",
    "trafilatura>=2.0.0",
]

[project.scripts]
fetch = "fetch:main"

[tool.uv]
package = true

[tool.setuptools]
packages = ["fetch"]

[tool.setuptools.dynamic]
version = {attr = "fetch.__version__.__version__"}

